기능:config 파일

+++ Drivers/Driver/driver_adc.h
@@ -45,8 +45,8 @@
 
 driver_t *driver_adc_open(uint32_t num,void *opt);
 void driver_close(driver_t *handle);
-int32_t driver_adc_single_read(driver_t *drv,int channel,uint16_t avg,uint8_t *err);
-int32_t driver_adc_diff_read(driver_t *drv,int channel,uint16_t avg,uint8_t *err);
+float driver_adc_single_read(driver_t *drv, int channel, uint16_t avg, uint8_t *err);
+float driver_adc_diff_read(driver_t *drv,int channel,uint16_t avg,uint8_t *err);
 void driver_adc_set(driver_t *handle, adc_set_option_t option, void *value);
 
 #endif
